feat: remove username from user
Can't figure out an easy way to allow users to edit their own name due
to how it's handled on the backend. It's implemented as a custom field
on the Account model, but Rodauth doesn't expose a path helper to change
the value, and I am not sure how to do validation for custom fields
using a hook like `before_account_update`. I'm sure if I had more chops
I could figure it out, but here we are... I also don't want to make it a
field that can't be changed, so it's removed.

Reverting back to using emails, I think it makes sense to simply omit a
user's name/ID when showing posts.