fix: update nav bar
I am not happy with what I've mapped out in Excalidraw, and I'm
considering two options. The first is to stick, more or less, with the
flow I've designed on Excalidraw and create a "user badge" component
that just stylizes the user name in the style of Google's first capital
letter. The other option is to build that component, but turn it into a dropdown
menu element which I can then do a little more with.

I'm leaning towards the former, as the truth is that I feel a strong
urge to move on from this project to pursue some other things, and I
think building a less fully featured component would benefit me in terms
of moving on.

Another option is to use some package to do the dropdown menu effect.