fix: validate additional attachments
The previous fix that removed `reject` caused Rails to replace attachments
rather than adding them as was the intended behaviour. Using `reject` means
that whatever is in the rejected keys is not validated as part of the `update`
call. This posed a problem, because we can't be sure the frontend client will
send the correct mime types (the frontend `accepts` is a soft restriction) so
we must also validate files on the server side (at least checking the mime
type). Because we must use `reject` to _append_ files to the
`has_many_attached` association, we were not getting file validation on
additional attachments.

My original approach was to add `before_update :image_types` in the model,
which will force the validation to run before the update, but the problem was
that it was only validating the model before the update. In otherwords, kind
of useless.

The solution seemed to end up involving skipping attaching files if they
didn't fit the right mimetype inside the controller. In order to
faciliate this, I added a `constants.rb` so that we could consolidate
the accepted types into one place and use it in both the model and the
controller.

With this approach, invalid files are catalogued in an `update_errors`
JSON field after the update, but are then forgotten about. I like this
approach because it's a silent error but we can update the frontend to
show this if we want in the future.