fix: move action bar to top
Also update DaisyUI to 2.50.2 in hopes of fixing a color flicker.

Sometimes, when you hover between the dropdown menu for the user and
the user badge, there is a flicker of color which blocks parts of the
user badge and even the "New Post" button.

The issue seems to be specific to Firefox, as I cannot seem to reproduce
it on my local Chrome browser. Interestingly enough, the flash of color
seems to take on whatever color is set for the underlying ``<article>`
border color - in the case of the GIF I captured, it was
`dark:border-accent/25` and did change once I removed that color
specification from the `<article>` element.

I've submitted the issue to daisyUI, though I suspect that this is a
problem with Firefox's rendering engine.